+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## NEXT
 ### Features
-* Polyline edtior supports drag shape events .
+* Polyline edtior supports drag shape events.
+* Add ability to change width to Hippodrome editor
 
 ## Fixes
+* Hippodrome editor - setManually supports positions and hippodrome width.
+* Fixed Polygons Editing  bug that happened when dragging a point.  
 * Hippodrome editor - setManually supports positions and hippodrome width
 
 ## Breaking changes